iOS rename DittoCombineExample to CombineProducts; move to SwiftUI-Combine
- rename DittoCombineExample to CombineProducts for clarity
- move into samples/SwiftUI-Combine directory for organization
- update SDK to v4.0.1
