--- DittoCombineExample/DittoCombineExample/DittoCombineExampleApp.swift
@@ -1,14 +1,0 @@
-import SwiftUI
-//import 
-
-@main
-struct DittoCombineExampleApp: App {
-    var body: some Scene {
-        WindowGroup {
-            NavigationView {
-                ProductsListView()
-            }
-            .navigationViewStyle(StackNavigationViewStyle())
-        }
-    }
-}
