--- DittoCombineExample/DittoCombineExample/StaticKeys.swift
@@ -1,33 +1,0 @@
-//
-//  StaticKeys.swift
-//  DittoCombineExample
-//
-//  Created by Eric Turner on 12/20/22.
-//
-
-import Foundation
-
-// MARK: Collection Keys
-let categoriesKey = "categories"
-let productsKey = "products"
-
-// MARK: Model Keys
-let dbIdKey = "_id"
-let idKey = "id"
-let nameKey = "name"
-let categoryIdKey = "categoryId"
-
-// MARK: UI Keys
-let productNameTitleKey = "Product Name"
-let editProductTitleKey = "Edit Product"
-let createProductTitleKey = "Create Product"
-
-let categorySectionKey = "Category"
-let categoriesTitleKey = "Categories"
-let addCategoryTitleKey = "Add Category"
-
-let circleFillImgKey = "circle.fill"
-let circleOpenImgKey = "circle"
-let requiredKey = "required"
-let saveChangesTitleKey = "Save Changes"
-let cancelTitleKey = "Cancel"
