--- SwiftUI-Combine/DittoCombineExample/DittoCombineExample.xcodeproj/project.xcworkspace/xcshareddata/swiftpm/Package.resolved
@@ -1,14 +1,0 @@
-{
-  "pins" : [
-    {
-      "identity" : "dittoswiftpackage",
-      "kind" : "remoteSourceControl",
-      "location" : "https://github.com/getditto/DittoSwiftPackage",
-      "state" : {
-        "revision" : "0f41078a6a7441cc56c6be952fa6bcd85b4e0199",
-        "version" : "3.0.0"
-      }
-    }
-  ],
-  "version" : 2
-}
