+++ SwiftUI-Combine/combine-products/CombineProducts.xcodeproj/project.xcworkspace/xcshareddata/swiftpm/Package.resolved
