+++ SwiftUI-Combine/combine-products/CombineProducts/CombineProductsApp.swift
@@ -1,0 +1,13 @@
+import SwiftUI
+
+@main
+struct CombineProductsApp: App {
+    var body: some Scene {
+        WindowGroup {
+            NavigationView {
+                ProductsListView()
+            }
+            .navigationViewStyle(StackNavigationViewStyle())
+        }
+    }
+}
